crypto-python/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.3.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.9
cryptography==3.4.8
dateparser==1.1.0
decorator==5.0.9
diagrams==0.20.0
distlib==0.3.2
filelock==3.0.12
frozenlist==1.2.0
graphviz==0.16
html5lib==1.1
idna==3.3
Jinja2==2.11.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.22.0
pickleDB==0.9.2
pipenv==2021.5.29
platformdirs==2.2.0
pycparser==2.20
python-binance==1.0.15
python-dateutil==2.8.2
python-telegram-bot==12.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.6
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
yarl==1.7.2
